requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0,<1.25.0
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0

# Web scraping
feedparser>=6.0.0
newspaper3k>=0.2.8
python-dateutil>=2.8.0

# Machine Learning - CPU optimized
scikit-learn>=1.0.0
transformers>=4.20.0
torch>=1.11.0
sentence-transformers>=2.0.0

# NLP libraries
nltk>=3.6
textblob>=0.15.0

# Web framework
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0

# Database
sqlalchemy>=1.4.0

# Data versioning
dvc>=2.10.0

# Scheduling moved to GitHub Actions - no schedule dependency needed

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
tqdm>=4.60.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Caching
joblib>=1.0.0